backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-Mail==0.9.1
pymysql
cryptography==41.0.3
python-dotenv==1.0.0
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.0.0
